Moved transformCandleSticks to helper folder.
